not so sure, again...

+++ src/V2D.cpp
@@ -6,7 +6,7 @@
 V2D::V2D()
 : x(0), y(0) {}
 
-V2D::V2D(double x, double y)
+V2D::V2D(float x, float y)
 : x(x), y(y) {}
 
 V2D::V2D(  V2D const& v) {
@@ -16,26 +16,26 @@ V2D::V2D(  V2D const& v) {
 
 //get&set methods -------------------------------------
 
-void V2D::getCoords(double& x, double& y) const
+void V2D::getCoords(float& x, float& y) const
 {
 	x = this->x;
 	y = this->y;
 }
 
-double V2D::getX() const
+float V2D::getX() const
 {
 	return this->x;
 }
 
-double V2D::getY() const
+float V2D::getY() const
 {
 	return this->y;
 }
 
 
-double V2D::getNorm() const
+float V2D::getNorm() const
 {
-	double norm(x*x + y*y);
+	float norm(x*x + y*y);
 	return norm;
 }
 
@@ -44,17 +44,17 @@ V2D V2D::normalize()
 	return (*this)/sqrt(getNorm());
 }
 
-void V2D::setCoords(double x, double y)
+void V2D::setCoords(float x, float y)
 {
 	this->x = x;
 	this->y = y;
 }
 
-void V2D::setX(double x){
+void V2D::setX(float x){
 	this->x = x;
 }
 
-void V2D::setY(double y){
+void V2D::setY(float y){
 	
 	this->y = y;
 }
@@ -63,7 +63,7 @@ void V2D::setY(double y){
 // internal operators
 V2D V2D::operator+(const V2D& v) const
 {
-	double x,y;
+	float x,y;
 	v.getCoords(x,y);
 	return V2D( (this->x+x),(this->y+y) );
 }
@@ -71,25 +71,25 @@ V2D V2D::operator+(const V2D& v) const
 
 V2D V2D::operator-(const V2D& v) const
 {
-	double x,y;
+	float x,y;
 	v.getCoords(x,y);
 	
 	return V2D( (this->x-x),(this->y-y) );
 }
 
-V2D V2D::operator*(const double& d) const
+V2D V2D::operator*(const float& d) const
 {
 	return V2D( (x*d),(y*d) );
 }
 
-V2D V2D::operator/(const double& d) const
+V2D V2D::operator/(const float& d) const
 {
 	return V2D( (x/d),(y/d) );
 }
 
-double V2D::operator|(const V2D& v) const
+float V2D::operator|(const V2D& v) const
 {
-	double x,y;
+	float x,y;
 	v.getCoords(x,y);
 	
 	return ( this->x*x + this->y*y );
@@ -104,7 +104,7 @@ V2D V2D::operator-(void) const
 
 void V2D::operator+=(const V2D& v)
 {
-	double x,y;
+	float x,y;
 	v.getCoords(x,y);
 	this->x += x;
 	this->y += y;
@@ -112,19 +112,19 @@ void V2D::operator+=(const V2D& v)
 
 void V2D::operator-=(const V2D& v)
 {
-	double x,y;
+	float x,y;
 	v.getCoords(x,y);
 	this->x -= x;
 	this->y -= y;
 }
 
-void V2D::operator*=(const double d)
+void V2D::operator*=(const float d)
 {
 	x *= d;
 	y *= d;
 }
 
-void V2D::operator/=(const double d)
+void V2D::operator/=(const float d)
 {
 	x /= d;
 	y /= d;
@@ -134,7 +134,7 @@ void V2D::operator/=(const double d)
 
 /*---------------------------------------------------------------------------*/
 // external operators
-V2D operator*(const double& d, const V2D& v)
+V2D operator*(const float& d, const V2D& v)
 {
 	return (v * d);
 }
